Filtered Array But the display can not be the same like design in figma need help

+++ exzo/Application/Product Page/View/ProductView.swift
@@ -19,16 +19,18 @@ struct ProductView: View {
                 }
                 Spacer()
                     .frame(height: 5)
-                List {
-                    ForEach($viewModel.products, id: \.self) { product in
-                        ZStack {
-                            NavigationLink {
-                                DetailProductView(product: product.wrappedValue)
-                            } label: {
+                if !viewModel.products.isEmpty {
+                    List {
+                        ForEach($viewModel.products, id: \.self) { product in
+                            ZStack {
+                                NavigationLink {
+                                    DetailProductView(product: product.wrappedValue)
+                                } label: {
+                                    ProductRow(product: product.wrappedValue)
+                                }
+                                .buttonStyle(PlainButtonStyle())
                                 ProductRow(product: product.wrappedValue)
                             }
-                            .buttonStyle(PlainButtonStyle())
-                            ProductRow(product: product.wrappedValue)
                         }
                     }
                     .onDelete(perform: viewModel.deleteItem)
